+++ src/main/java/faang/school/projectservice/controller/project/ProjectController.java
@@ -2,11 +2,11 @@ package faang.school.projectservice.controller.project;
 
 import faang.school.projectservice.dto.project.ProjectDto;
 import faang.school.projectservice.dto.project.ProjectFilterDto;
-import faang.school.projectservice.service.project.ProjectService;
-import jakarta.validation.Valid;
+import faang.school.projectservice.service.ProjectService;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,31 +25,31 @@ public class ProjectController {
 
     private final ProjectService projectService;
 
-    @PostMapping("{userId}/create")
+    @PostMapping("/{userId}")
     @ResponseStatus(HttpStatus.CREATED)
-    public ProjectDto create(@Valid @PathVariable Long userId, @Valid @RequestBody ProjectDto projectDto) {
+    public ProjectDto create(@PathVariable Long userId, @Validated @RequestBody ProjectDto projectDto) {
         return projectService.create(userId, projectDto);
     }
 
-    @PutMapping("/update")
-    public ProjectDto update(@Valid @RequestBody ProjectDto projectDto) {
+    @PutMapping("/")
+    public ProjectDto update(@Validated @RequestBody ProjectDto projectDto) {
         return projectService.update(projectDto);
     }
 
     @PostMapping("/{userId}/accessible/filtered")
-    public List<ProjectDto> findProjectsWithFilter(@Valid @PathVariable Long userId,
-                                                   @Valid @NotNull @RequestBody ProjectFilterDto projectFilterDto) {
+    public List<ProjectDto> findProjectsWithFilter(@PathVariable Long userId,
+                                                   @Validated @NotNull @RequestBody ProjectFilterDto projectFilterDto) {
         return projectService.findProjectsWithFilter(userId, projectFilterDto);
     }
 
     @GetMapping("/{userId}/accessible/projects")
-    public List<ProjectDto> findAllProjects(@Valid @PathVariable Long userId) {
+    public List<ProjectDto> findAllProjects(@PathVariable Long userId) {
         return projectService.findAllProjects(userId);
     }
 
     @PostMapping("/{userId}/accessible/project")
-    public  ProjectDto findById(@Valid @PathVariable Long userId,
-                                @Valid @RequestBody ProjectFilterDto projectFilterDto) {
+    public  ProjectDto findById(@PathVariable Long userId,
+                                @Validated @RequestBody ProjectFilterDto projectFilterDto) {
         return projectService.findById(userId, projectFilterDto);
     }
 }
