--- src/main/java/faang/school/projectservice/service/project/ProjectService.java
@@ -1,99 +1,0 @@
-package faang.school.projectservice.service.project;
-
-import faang.school.projectservice.dto.project.ProjectDto;
-import faang.school.projectservice.dto.project.ProjectFilterDto;
-import faang.school.projectservice.filter.project.ProjectFilter;
-import faang.school.projectservice.mapper.project.ProjectMapper;
-import faang.school.projectservice.model.Project;
-import faang.school.projectservice.model.ProjectStatus;
-import faang.school.projectservice.model.ProjectVisibility;
-import faang.school.projectservice.repository.ProjectRepository;
-import faang.school.projectservice.service.project.validation.ProjectServiceValidator;
-import jakarta.persistence.EntityNotFoundException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class ProjectService {
-
-    private final ProjectServiceValidator projectServiceValidator;
-    private final ProjectMapper projectMapper;
-    private final ProjectRepository projectRepository;
-    private final List<ProjectFilter> filters;
-
-    @Transactional
-    public ProjectDto create(Long userId, ProjectDto projectDto) {
-        projectServiceValidator.validateCreation(userId, projectDto);
-        Project project = projectMapper.toEntity(projectDto);
-        entityCreatePreparation(userId, project);
-        project = projectRepository.save(project);
-        return projectMapper.toDto(project);
-    }
-
-    @Transactional
-    public ProjectDto update(ProjectDto projectDto) {
-        projectServiceValidator.validateUpdating(projectDto);
-        Project project = projectMapper.toEntity(projectDto);
-        entityUpdatePreparation(project);
-        project = projectRepository.save(project);
-        return projectMapper.toDto(project);
-    }
-
-    @Transactional(readOnly = true)
-    public List<ProjectDto> findProjectsWithFilter(Long userId, ProjectFilterDto projectFilterDto) {
-        List<ProjectFilter> applicableFilters = getApplicableProjectFilters(projectFilterDto);
-        List<Project> filteredProjects = projectRepository.findAllAvailableProjectsByUserId(userId).toList();
-        for (ProjectFilter applicableFilter : applicableFilters) {
-            filteredProjects = applicableFilter.apply(filteredProjects.stream(), projectFilterDto).toList();
-        }
-        return projectMapper.toDtos(filteredProjects);
-    }
-
-    @Transactional(readOnly = true)
-    public List<ProjectDto> findAllProjects(Long userId) {
-        return projectMapper.toDtos(projectRepository
-                .findAllAvailableProjectsByUserId(userId)
-                .toList());
-    }
-
-    @Transactional(readOnly = true)
-    public ProjectDto findById(Long userId, ProjectFilterDto projectFilterDto) {
-        return projectMapper.toDto(projectRepository
-                .findAvailableByUserIdAndProjectId(userId, projectFilterDto.getIdPattern())
-                .orElseThrow(() -> {
-                    log.info("Project with id {} not found", projectFilterDto.getIdPattern());
-                    return new EntityNotFoundException("Project not found");
-                }));
-    }
-
-    private void entityCreatePreparation(Long userId, Project project) {
-        project.setOwnerId(userId);
-        project.setStatus(ProjectStatus.CREATED);
-        if (project.getVisibility() == null) {
-            project.setVisibility(ProjectVisibility.PUBLIC);
-        }
-    }
-
-    private void entityUpdatePreparation(Project project) {
-        if (checkUpdateReason(project)) {
-            project.setUpdatedAt(LocalDateTime.now());
-        }
-    }
-
-    private boolean checkUpdateReason(Project project) {
-        return (project.getDescription() != null || project.getStatus() != null);
-    }
-
-    private List<ProjectFilter> getApplicableProjectFilters(ProjectFilterDto projectFilterDto) {
-        return filters.stream()
-                .filter(filter -> filter.isApplicable(projectFilterDto))
-                .toList();
-    }
-}
